File uploader with display of files
Requires deploying the code to every site directory in
production...unless we can handle it a little differently.  Tried using
relative paths with the site session name, the $GLOBALS['OE_SITE_DIR'],
$OE_SITE_DIR, $webroot, and a lot of other ways to no avail.  This is
possibly due to the archaic ?is_windows usage in globals.php, sandboxing
in the .js or a million other things.  Placing the code directly in the
sites directory and using the session site id might have worked, except
for the fact that it might mess with the multisite functions to add
directories there...Hair pulling out is over, and this works...ready for
review or help with the damn pathing.