Zipping was Zagging
After banging my head against the wall, I figured out that somehow the
system was caching the old run of reports and writing them to the zip
file.  I ran the new code again and again, deleting the previous run of
files, and the modal window showed that the new code was running, but
again and again the multiple files generated by the original XML export
file for PQRS were getting created in the zip.
Finally, I removed the file moving and writing code, and a single file
with all the measures appeared.  This is a puzzle for a Linux Sysadmin.