insert HCC status updates
Next element for HCC updates
insert HCC status into the patient issues.  May be some platform issues with the use of REPLACE syntax if table keys are not tweaked.  'lists' table may be abandoned with this, but this is the easiest way to get a dems window status.  lists table sucks as a vertical table due to key schema.