Fee sheet functions
Pulled out the non-js fee sheet functions into an include to shorten the mess a bit.
Changed the drop-down menus to print the code and modifiers in front of the name so that the codes list text need not be modified to include the code itself.  This will require some superbill (administration>codes) edits for folks that have already entered descriptions including the codes.