user admin changes
Also corrected missing fontawesome link.  Might not be a version that works...
Also, the calendar mouseover event for showing the patient profile picture is dorked up.  I changed the url to reflect the /sites/ directory location, and that looks OK in the output code, but image still does not render (but no broken link).